requirements.txt
anyio==4.4.0
httpx==0.27.0
urllib3==2.2.2
uvicorn==0.30.1
asgiref==3.7.2
Flask==3.0.3
httpcore==1.0.5
requests==2.32.3
python-dotenv==1.0.1
python-telegram-bot==20.8
mysql-connector-python==8.0.33
#blinker==1.8.2
#certifi==2024.6.2
#cffi==1.16.0
#charset-normalizer==3.3.2
#click==8.1.7
#cssselect2==0.7.0
#defusedxml==0.7.1
#exceptiongroup==1.2.1
#h11==0.14.0
#idna==3.7
#importlib_metadata==7.1.0
#itsdangerous==2.2.0
#Jinja2==3.1.4
#MarkupSafe==2.1.5
#pillow==10.3.0
#psutil==6.0.0
#pycparser==2.22
#sniffio==1.3.1
#tinycss2==1.3.0
#typing_extensions==4.12.2
#webencodings==0.5.1
#Werkzeug==3.0.3
#zipp==3.19.2
